Bogus Code Done

+++ EvaluationSystemServer/EvaluationSystemServer/DataModels/RequestModels/Certificates/CertificateRequestModel.cs
@@ -17,7 +17,7 @@
         /// <summary>
         /// The graduation year
         /// </summary>
-        public double? GradutationYear { get; set; }
+        public int? GradutationYear { get; set; }
 
         /// <summary>
         /// The grade
